I changed it by accident

+++ EmulatorMS6502/CPU/MOS6502.cs
@@ -11,11 +11,8 @@ namespace EmulatorMOS6502.CPU {
 
         // Trzyma miejsce w pamięci z którego obecnie odczytujemy
         UInt16 abs_address = 0x0000;
-        Byte rel_address = 0x00;
+        UInt16 rel_address = 0x00;
         Byte opcode = 0x00;
-
-   
-   
 
        // 1. parita
         #region flags
